Add Backstage Plugin Gallery
Add a plugin gallery to `/backstage/plugins` which lists Rollbar and
Sentry (just to get started). Clicking each plugin brings the user to a
page which explains how to install and configure the plugin. The
plugins are described as a combination of YAML file and a MD file
which has arbitrary notes about the plugin.

 - Add logos into `content/assets`.
 - add a custom YAML transfomer plugin for loading plugins YAMLs and
   replacing image filesystem paths with sharp image objects.
 - Abstract some components to make them useful on this new page.