Upgrade various dependencies
The aim was to enable lazy generation of issues in order to
speed up dev server startup times. They had grown to 2+ mins
because of all the plugin logos.

Contains the following upgrades:

 - babel-preset-gatsby
 - gatsby
 - gatsby-image
 - gatsby-plugin-sharp
 - gatsby-plugin-manifest
 - gatsby-remark-images
 - gatsby-plugin-sitemap

I've also added @babel/core to solve unmet dependency warnings.

I hit an issue where `yarn build` would work locally on OSX and fail on
GitHub actions and on Netlify. It turned out to be this problem:
https://github.com/gatsbyjs/gatsby/issues/21032