Add sitemap and slash variants to robots.txt
The privacy and terms pages are disallowed because I don't want them
cached, incase someone misses a change.

I've noticed that Google Analytics is recording both trailing slash and
non-trailing slash URLs. This change ensures that both versions are
excluded.

It also adds a link to the sitemap. Not really necessary, as Google
knows it already, but no harm.