Add legal notices
This commit adds some legal notices to our website in order to help
legal teams understand what data we collect and what we use it for. It
collects all of these notices under one path `/legal-notices` so they
are all easily available in one place.

 - Add Acceptable use Policy
 - Add Cookies Policy
 - Add list of sub-processors
 - Move Terms and Privacy policy to new legal-notices area.
 - Set up 301 redirects for the old `/terms/` and `/privacy/` pages.

There are also two technical updates/refactors in this commit.

 1. I abstracted some logic for creating Gatsby pages in
    `gatsby-node.js`. There was lots of duplicate logic.
 2. I updated the caniuse database. caniuse is a tool which is used to
    target compatibility with certain browser segments like "last 2
    chrome versions".