Add docs
Adds the `/docs/` sub-path to the website where documentation can be
stored. So far, all that is written is a few integration guides and a
Getting Started for Admins guide.

Existing Blog posts are affected by this change because there seems
to be no way to have images left aligned on some pages and center
aligned on others. There's no way to remove the default styling
added by gatsby-remark-images and then reset it.

I've left aligned all images because it looks really bad in the docs
when images are left aligned but looks ok in the blog posts when images
are not center aligned (as they previously were).