Simplify use of gatsby-source-filesystem
I guess I didn't understand how this was supposed to work when I first
started using Gatsby. I had one instance of gatsby-source-filesystem for
each subdirectory of `content`. This would create slugs which didn't
have the correct prefix on them (like `blog/xyz`). I dealt with this by
manually adding the `/blog/` prefix to each URL in `gatsby-node.js`.

A much simpler way is to load all subdirectories of content with one
instance of gatsby-source-filesystem. Then you get the directory names
as part of the URL automatically and the manual prefixing can be removed
from `gatsby-node.js`.

This change is non-user facing. The public URLs of the website should
all remain the same as they are now.