Use frontmatter for plugin descriptions
Previously, 2 different directories stored the data which was used to
render the plugin marketplace. There was `content/plugins/descriptions`
and `content/plugins/notes`. The descriptions directory had YAML files
which stored the name and logo of each plugin and the notes which had
additional information about what the plugin would be used for.

This commit merges these two functions together inside the
`content/backstage/plugins` directory. Each plugin is a single markdown
file. The name, logo and other standardized info is in the frontmatter
and the additional info is in the body of the markdown file.

The benefits are:

 1. It will allow me to upgrade the sharp library and fix #105.
    Currently blocked because no YAML transformer for gatsby will
    auto-transform paths in YAML files into images. By using markdown
    frontmatter instead I get this feature automatically.
 2. Allows me to remove the custom transformer in the plugins
    directory. This is untested and hacky code which I can't maintain.
 3. It's simpler to just have one file per plugin.