Use <StickyFooter /> in more places
I believe we were opting not to use the StickyFooter on these pages
previously because it didn't have an many layout features as it does
now.

The edited pages had two LayoutControl elements involved in rendering
the SitewideHeader and SitewideFooter which was causing extra horizontal
padding and some unexpected text reflowing.

I have checked the modals and they still work in this setup.