Force a trailing slash on internal links
This is a continuation of https://github.com/RoadieHQ/marketing-site/pull/94,
or of `9923f1df2957370afe70a303d101af289e46d4bd` more specifically.

The plugin added in that PR was only part of the requirements. We also
need to do step 3 described on [this
page](https://www.gatsbyjs.com/plugins/gatsby-plugin-force-trailing-slashes/?=Link).

> All uses of Gatsby’s <Link> component in your code to intentionally
> specify a trailing slash.

This commit does that.