Remove dotenv
Dotenv had started failing the build on Netlify and Gatsby because of
[this error](https://github.com/motdotla/dotenv/issues/233).

Given the choice between adding a custom webpack config to gatsby simply
so I could go against the wishes of the dotenv maintainers, or removing
dotenv entirely, I chose the latter.

Dotenv was only used in local development to load some algolia secrets
from a `.env` file. It's feature can be replaced with a simple change in
the way we run Gatsby:

```
env $(cat .env | xargs) yarn build
```