Copy Roadie R to static
I'm customizing the Auth0 sign in page and we need to add the URL of a
logo. The logo already exists in the `content/assets` directory but
the URLs of those assets are dynamically generated based on a hash of
the content. We need a static URL. Placing the logo in the `static`
directory achieves that.