Send tracking info to GA, plausible and Netlify
This will allow us to use Netlify Split Testing and track the results in
our analytics tools. This will send x% of traffic to deployed git branch
A and 1-x% of traffic to deployed git branch B.

We send additional data to analytics tools so we can determine which
branch got more conversions.

 1. The name of the deployed branch is sent to Google Analytics as
    Dimension Data.
 2. The name of the deployed branch is sent to plausible as custom data.
 3. The clicked button label and the name of the deployed branch is sent
    to netlify forms.