Remove reference to manual / CI triggering of custom pipeline
I propose we remove references to users being able to programatically trigger the custom plugins pipeline. We don't have any users doing this but our documentation has these options as #1, #2 which seems to encourage this. 
While we certainly can do this I think for now we should keep it internal as a workaround if a user needs it. It's not ideal as the we have to 1) supply / handover the token manually 2) the webhook url is an AWS url which has changed in the past 3) document the request they need to send. All of these seem like internal details to me.