Create a flow to the custom plugins docs
Add links to the other custom plugins docs so this can be used as the "entry point" to custom plugins. The path of least resistance is where the user configures the plugin before they trigger a build, that way the first version is built into the application (while this might not be the intuitive order given how the feature works now I think this is best?).